requirements.txt
requests==2.32.3
SQLAlchemy==2.0.31
schedule==1.2.2  
python-dotenv==1.0.1